Renamed VirtualFlowContent & VirtualFlow to represent scroll bars added or not: now named VirtualFlow & VirtualizedScrollPane, respectively.
