+++ src/main/java/org/fxmisc/flowless/VirtualFlow.java
@@ -1,172 +1,406 @@
 package org.fxmisc.flowless;
 
-import javafx.application.Platform;
-import javafx.beans.binding.Bindings;
-import javafx.beans.binding.DoubleBinding;
+import java.util.Optional;
+import java.util.function.Function;
+
+import javafx.beans.property.ReadOnlyDoubleProperty;
+import javafx.beans.property.ReadOnlyDoubleWrapper;
+import javafx.collections.ObservableList;
 import javafx.geometry.Bounds;
 import javafx.geometry.Orientation;
-import javafx.scene.Node;
-import javafx.scene.control.ScrollBar;
+import javafx.scene.input.ScrollEvent;
 import javafx.scene.layout.Region;
+import javafx.scene.shape.Rectangle;
 
+import org.reactfx.collection.MemoizationList;
+import org.reactfx.util.Lists;
 import org.reactfx.value.Val;
 import org.reactfx.value.Var;
 
-public class VirtualFlow<V extends Node & Virtualized> extends Region implements Virtualized {
+class VirtualFlow<T, C extends Cell<T, ?>> extends Region implements Virtualized {
 
-    private final ScrollBar hbar;
-    private final ScrollBar vbar;
-    private final V content;
+    public static enum Gravity { FRONT, REAR }
 
-    public VirtualFlow(V content) {
-        this.getStyleClass().add("virtual-flow");
-        this.content = content;
-
-        // create scrollbars
-        hbar = new ScrollBar();
-        vbar = new ScrollBar();
-        hbar.setOrientation(Orientation.HORIZONTAL);
-        vbar.setOrientation(Orientation.VERTICAL);
-
-        // scrollbar ranges
-        hbar.setMin(0);
-        vbar.setMin(0);
-        hbar.maxProperty().bind(content.totalWidthEstimateProperty());
-        vbar.maxProperty().bind(content.totalHeightEstimateProperty());
-
-        // scrollbar increments
-        setupUnitIncrement(hbar);
-        setupUnitIncrement(vbar);
-        hbar.blockIncrementProperty().bind(hbar.visibleAmountProperty());
-        vbar.blockIncrementProperty().bind(vbar.visibleAmountProperty());
-
-        // scrollbar positions
-        Bindings.bindBidirectional(
-                Var.doubleVar(hbar.valueProperty()),
-                content.estimatedScrollXProperty());
-        Bindings.bindBidirectional(
-                Var.doubleVar(vbar.valueProperty()),
-                content.estimatedScrollYProperty());
-
-        // scrollbar visibility
-        Val<Double> layoutWidth = Val.map(layoutBoundsProperty(), Bounds::getWidth);
-        Val<Double> layoutHeight = Val.map(layoutBoundsProperty(), Bounds::getHeight);
-        Val<Boolean> needsHBar0 = Val.combine(
-                content.totalWidthEstimateProperty(),
-                layoutWidth,
-                (cw, lw) -> cw > lw);
-        Val<Boolean> needsVBar0 = Val.combine(
-                content.totalHeightEstimateProperty(),
-                layoutHeight,
-                (ch, lh) -> ch > lh);
-        Val<Boolean> needsHBar = Val.combine(
-                needsHBar0,
-                needsVBar0,
-                content.totalWidthEstimateProperty(),
-                vbar.widthProperty(),
-                layoutWidth,
-                (needsH, needsV, cw, vbw, lw) -> needsH || needsV && cw + vbw.doubleValue() > lw);
-        Val<Boolean> needsVBar = Val.combine(
-                needsVBar0,
-                needsHBar0,
-                content.totalHeightEstimateProperty(),
-                hbar.heightProperty(),
-                layoutHeight,
-                (needsV, needsH, ch, hbh, lh) -> needsV || needsH && ch + hbh.doubleValue() > lh);
-        hbar.visibleProperty().bind(needsHBar);
-        vbar.visibleProperty().bind(needsVBar);
-
-        // request layout later, because if currently in layout, the request is ignored
-        hbar.visibleProperty().addListener(obs -> Platform.runLater(() -> requestLayout()));
-        vbar.visibleProperty().addListener(obs -> Platform.runLater(() -> requestLayout()));
-
-        getChildren().addAll(content, hbar, vbar);
+    public static <T, C extends Cell<T, ?>> VirtualFlow<T, C> createHorizontal(
+            ObservableList<T> items,
+            Function<? super T, ? extends C> cellFactory) {
+        return createHorizontal(items, cellFactory, Gravity.FRONT);
     }
 
-    public Val<Double> totalWidthEstimateProperty() {
-        return content.totalWidthEstimateProperty();
+    public static <T, C extends Cell<T, ?>> VirtualFlow<T, C> createHorizontal(
+            ObservableList<T> items,
+            Function<? super T, ? extends C> cellFactory,
+            Gravity gravity) {
+        return new VirtualFlow<>(items, cellFactory, new HorizontalHelper(), gravity);
     }
 
-    public Val<Double> totalHeightEstimateProperty() {
-        return content.totalHeightEstimateProperty();
+    public static <T, C extends Cell<T, ?>> VirtualFlow<T, C> createVertical(
+            ObservableList<T> items,
+            Function<? super T, ? extends C> cellFactory) {
+        return createVertical(items, cellFactory, Gravity.FRONT);
     }
 
-    public Var<Double> estimatedScrollXProperty() {
-        return content.estimatedScrollXProperty();
+    public static <T, C extends Cell<T, ?>> VirtualFlow<T, C> createVertical(
+            ObservableList<T> items,
+            Function<? super T, ? extends C> cellFactory,
+            Gravity gravity) {
+        return new VirtualFlow<>(items, cellFactory, new VerticalHelper(), gravity);
     }
 
-    public Var<Double> estimatedScrollYProperty() {
-        return content.estimatedScrollYProperty();
+    private final ObservableList<T> items;
+    private final OrientationHelper orientation;
+    private final CellListManager<T, C> cellListManager;
+    private final SizeTracker sizeTracker;
+    private final CellPositioner<T, C> cellPositioner;
+    private final Navigator<T, C> navigator;
+
+    // non-negative
+    private final ReadOnlyDoubleWrapper breadthOffset = new ReadOnlyDoubleWrapper(0.0);
+    public ReadOnlyDoubleProperty breadthOffsetProperty() {
+        return breadthOffset.getReadOnlyProperty();
     }
 
-    @Override
-    protected double computePrefWidth(double height) {
-        return content.prefWidth(height);
+    public Val<Double> totalBreadthEstimateProperty() {
+        return sizeTracker.maxCellBreadthProperty();
     }
 
-    @Override
-    protected double computePrefHeight(double width) {
-        return content.prefHeight(width);
+    private final Val<Double> breadthPositionEstimate;
+    public Var<Double> breadthPositionEstimateProperty() {
+        return breadthPositionEstimate.asVar(this::setBreadthPosition);
     }
 
-    @Override
-    protected double computeMinWidth(double height) {
-        return vbar.minWidth(-1);
+    private final Val<Double> lengthOffsetEstimate;
+
+    private final Val<Double> lengthPositionEstimate;
+    public Var<Double> lengthPositionEstimateProperty() {
+        return lengthPositionEstimate.asVar(this::setLengthPosition);
     }
 
-    @Override
-    protected double computeMinHeight(double width) {
-        return hbar.minHeight(-1);
+    VirtualFlow(
+            ObservableList<T> items,
+            Function<? super T, ? extends C> cellFactory,
+            OrientationHelper orientation,
+            Gravity gravity) {
+        this.getStyleClass().add("virtual-flow-content");
+        this.items = items;
+        this.orientation = orientation;
+        this.cellListManager = new CellListManager<>(items, cellFactory);
+        MemoizationList<C> cells = cellListManager.getLazyCellList();
+        this.sizeTracker = new SizeTracker(orientation, layoutBoundsProperty(), cells);
+        this.cellPositioner = new CellPositioner<>(cellListManager, orientation, sizeTracker);
+        this.navigator = new Navigator<>(cellListManager, cellPositioner, orientation, gravity, sizeTracker);
+
+        getChildren().add(navigator);
+        clipProperty().bind(Val.map(
+                layoutBoundsProperty(),
+                b -> new Rectangle(b.getWidth(), b.getHeight())));
+
+
+        // set up bindings
+
+        breadthPositionEstimate = Val.combine(
+                breadthOffset,
+                sizeTracker.viewportBreadthProperty(),
+                sizeTracker.maxCellBreadthProperty(),
+                (off, vpBr, totalBr) -> offsetToScrollbarPosition(off.doubleValue(), vpBr, totalBr));
+
+        lengthOffsetEstimate = sizeTracker.lengthOffsetEstimateProperty();
+
+        lengthPositionEstimate = Val.combine(
+                lengthOffsetEstimate,
+                sizeTracker.viewportLengthProperty(),
+                sizeTracker.totalLengthEstimateProperty(),
+                (off, vpLen, totalLen) -> offsetToScrollbarPosition(off, vpLen, totalLen))
+                .orElseConst(0.0);
+
+        // scroll content by mouse scroll
+        this.addEventHandler(ScrollEvent.SCROLL, se -> {
+            scrollXBy(-se.getDeltaX());
+            scrollYBy(-se.getDeltaY());
+            se.consume();
+        });
     }
 
-    @Override
-    protected double computeMaxWidth(double height) {
-        return content.maxWidth(height);
+    public void dispose() {
+        navigator.dispose();
+        sizeTracker.dispose();
+        cellListManager.dispose();
     }
 
-    @Override
-    protected double computeMaxHeight(double width) {
-        return content.maxHeight(width);
+    public C getCellFor(int itemIndex) {
+        Lists.checkIndex(itemIndex, items.size());
+        return cellPositioner.getSizedCell(itemIndex);
+    }
+
+    public Optional<C> getCellIfVisible(int itemIndex) {
+        return cellPositioner.getCellIfVisible(itemIndex);
+    }
+
+    public ObservableList<C> visibleCells() {
+        return cellListManager.getLazyCellList().memoizedItems();
+    }
+
+    public Val<Double> totalLengthEstimateProperty() {
+        return sizeTracker.totalLengthEstimateProperty();
     }
 
     @Override
     protected void layoutChildren() {
-        double layoutWidth = getLayoutBounds().getWidth();
-        double layoutHeight = getLayoutBounds().getHeight();
-        boolean vbarVisible = vbar.isVisible();
-        boolean hbarVisible = hbar.isVisible();
-        double vbarWidth = vbarVisible ? vbar.prefWidth(-1) : 0;
-        double hbarHeight = hbarVisible ? hbar.prefHeight(-1) : 0;
 
-        double w = layoutWidth - vbarWidth;
-        double h = layoutHeight - hbarHeight;
-
-        content.resize(w, h);
-
-        hbar.setVisibleAmount(w);
-        vbar.setVisibleAmount(h);
-
-        if(vbarVisible) {
-            vbar.resizeRelocate(layoutWidth - vbarWidth, 0, vbarWidth, h);
+        // navigate to the target position and fill viewport
+        while(true) {
+            double oldLayoutBreadth = sizeTracker.getCellLayoutBreadth();
+            orientation.resize(navigator, oldLayoutBreadth, sizeTracker.getViewportLength());
+            navigator.layout();
+            if(oldLayoutBreadth == sizeTracker.getCellLayoutBreadth()) {
+                break;
+            }
         }
 
-        if(hbarVisible) {
-            hbar.resizeRelocate(0, layoutHeight - hbarHeight, w, hbarHeight);
+        orientation.relocate(navigator, -breadthOffset.get(), 0);
+    }
+
+    @Override
+    protected final double computePrefWidth(double height) {
+        switch(getContentBias()) {
+            case HORIZONTAL: // vertical flow
+                return computePrefBreadth();
+            case VERTICAL: // horizontal flow
+                return computePrefLength(height);
+            default:
+                throw new AssertionError("Unreachable code");
         }
     }
 
-    private static void setupUnitIncrement(ScrollBar bar) {
-        bar.unitIncrementProperty().bind(new DoubleBinding() {
-            { bind(bar.maxProperty(), bar.visibleAmountProperty()); }
+    @Override
+    protected final double computePrefHeight(double width) {
+        switch(getContentBias()) {
+            case HORIZONTAL: // vertical flow
+                return computePrefLength(width);
+            case VERTICAL: // horizontal flow
+                return computePrefBreadth();
+            default:
+                throw new AssertionError("Unreachable code");
+        }
+    }
 
-            @Override
-            protected double computeValue() {
-                double max = bar.getMax();
-                double visible = bar.getVisibleAmount();
-                return max > visible
-                        ? 16 / (max - visible) * max
-                        : 0;
+    private double computePrefBreadth() {
+        return 100;
+    }
+
+    private double computePrefLength(double breadth) {
+        return 100;
+    }
+
+    @Override
+    public final Orientation getContentBias() {
+        return orientation.getContentBias();
+    }
+
+    void scrollLength(double deltaLength) {
+        setLengthOffset(lengthOffsetEstimate.getValue() + deltaLength);
+    }
+
+    void scrollBreadth(double deltaBreadth) {
+        setBreadthOffset(breadthOffset.get() + deltaBreadth);
+    }
+
+    void scrollXBy(double deltaX) {
+        orientation.scrollHorizontallyBy(this, deltaX);
+    }
+
+    void scrollYBy(double deltaY) {
+        orientation.scrollVerticallyBy(this, deltaY);
+    }
+
+    void scrollXToPixel(double pixel) {
+        orientation.scrollHorizontallyToPixel(this, pixel);
+    }
+
+    void scrollYToPixel(double pixel) {
+        orientation.scrollVerticallyToPixel(this, pixel);
+    }
+
+    public Val<Double> totalWidthEstimateProperty() {
+        return orientation.widthEstimateProperty(this);
+    }
+
+    public Val<Double> totalHeightEstimateProperty() {
+        return orientation.heightEstimateProperty(this);
+    }
+
+    public Var<Double> estimatedScrollXProperty() {
+        return orientation.horizontalPositionProperty(this);
+    }
+
+    public Var<Double> estimatedScrollYProperty() {
+        return orientation.verticalPositionProperty(this);
+    }
+
+    VirtualFlowHit<C> hit(double x, double y) {
+        double bOff = orientation.getX(x, y);
+        double lOff = orientation.getY(x, y);
+
+        bOff += breadthOffset.get();
+
+        if(items.isEmpty()) {
+            return orientation.hitAfterCells(bOff, lOff);
+        }
+
+        layout();
+
+        int firstVisible = cellPositioner.getFirstVisibleIndex().getAsInt();
+        firstVisible = navigator.fillBackwardFrom0(firstVisible, lOff);
+        C firstCell = cellPositioner.getVisibleCell(firstVisible);
+
+        int lastVisible = cellPositioner.getLastVisibleIndex().getAsInt();
+        lastVisible = navigator.fillForwardFrom0(lastVisible, lOff);
+        C lastCell = cellPositioner.getVisibleCell(lastVisible);
+
+        if(lOff < orientation.minY(firstCell)) {
+            return orientation.hitBeforeCells(bOff, lOff - orientation.minY(firstCell));
+        } else if(lOff >= orientation.maxY(lastCell)) {
+            return orientation.hitAfterCells(bOff, lOff - orientation.maxY(lastCell));
+        } else {
+            for(int i = firstVisible; i <= lastVisible; ++i) {
+                C cell = cellPositioner.getVisibleCell(i);
+                if(lOff < orientation.maxY(cell)) {
+                    return orientation.cellHit(i, cell, bOff, lOff - orientation.minY(cell));
+                }
             }
-        });
+            throw new AssertionError("unreachable code");
+        }
+    }
+
+    void show(double viewportOffset) {
+        if(viewportOffset < 0) {
+            navigator.scrollTargetPositionBy(viewportOffset);
+        } else if(viewportOffset > sizeTracker.getViewportLength()) {
+            navigator.scrollTargetPositionBy(viewportOffset - sizeTracker.getViewportLength());
+        } else {
+            // do nothing, offset already in the viewport
+        }
+    }
+
+    void show(int itemIdx) {
+        navigator.setTargetPosition(new MinDistanceTo(itemIdx));
+    }
+
+    void showAsFirst(int itemIdx) {
+        navigator.setTargetPosition(new StartOffStart(itemIdx, 0.0));
+    }
+
+    void showAsLast(int itemIdx) {
+        navigator.setTargetPosition(new EndOffEnd(itemIdx, 0.0));
+    }
+
+    void showAtOffset(int itemIdx, double offset) {
+        navigator.setTargetPosition(new StartOffStart(itemIdx, offset));
+    }
+
+    void showRegion(int itemIndex, Bounds region) {
+      navigator.showLengthRegion(itemIndex, orientation.minY(region), orientation.maxY(region));
+      showBreadthRegion(orientation.minX(region), orientation.maxX(region));
+    }
+
+    private void showBreadthRegion(double fromX, double toX) {
+        double bOff = breadthOffset.get();
+        double spaceBefore = fromX - bOff;
+        double spaceAfter = sizeTracker.getViewportBreadth() - toX + bOff;
+        if(spaceBefore < 0 && spaceAfter > 0) {
+            double shift = Math.min(-spaceBefore, spaceAfter);
+            setBreadthOffset(bOff - shift);
+        } else if(spaceAfter < 0 && spaceBefore > 0) {
+            double shift = Math.max(spaceAfter, -spaceBefore);
+            setBreadthOffset(bOff - shift);
+        }
+    }
+
+    private void setLengthPosition(double pos) {
+        setLengthOffset(lengthPositionToPixels(pos));
+    }
+
+    private void setBreadthPosition(double pos) {
+        setBreadthOffset(breadthPositionToPixels(pos));
+    }
+
+    void setLengthOffset(double pixels) {
+        double total = totalLengthEstimateProperty().getOrElse(0.0);
+        double length = sizeTracker.getViewportLength();
+        double max = Math.max(total - length, 0);
+        double current = lengthOffsetEstimate.getValue();
+
+        if(pixels > max) pixels = max;
+        if(pixels < 0) pixels = 0;
+
+        double diff = pixels - current;
+        if(diff == 0) {
+            // do nothing
+        } else if(Math.abs(diff) < length) { // distance less than one screen
+            navigator.scrollTargetPositionBy(diff);
+        } else {
+            jumpToAbsolutePosition(pixels);
+        }
+    }
+
+    void setBreadthOffset(double pixels) {
+        double total = totalBreadthEstimateProperty().getValue();
+        double breadth = sizeTracker.getViewportBreadth();
+        double max = Math.max(total - breadth, 0);
+        double current = breadthOffset.get();
+
+        if(pixels > max) pixels = max;
+        if(pixels < 0) pixels = 0;
+
+        if(pixels != current) {
+            breadthOffset.set(pixels);
+            requestLayout();
+            // TODO: could be safely relocated right away?
+            // (Does relocation request layout?)
+        }
+    }
+
+    private void jumpToAbsolutePosition(double pixels) {
+        if(items.isEmpty()) {
+            return;
+        }
+
+        // guess the first visible cell and its offset in the viewport
+        double avgLen = sizeTracker.getAverageLengthEstimate().orElse(0.0);
+        if(avgLen == 0.0) return;
+        int first = (int) Math.floor(pixels / avgLen);
+        double firstOffset = -(pixels % avgLen);
+
+        if(first < items.size()) {
+            navigator.setTargetPosition(new StartOffStart(first, firstOffset));
+        } else {
+            navigator.setTargetPosition(new EndOffEnd(items.size() - 1, 0.0));
+        }
+    }
+
+    private double lengthPositionToPixels(double pos) {
+        double total = totalLengthEstimateProperty().getOrElse(0.0);
+        double length = sizeTracker.getViewportLength();
+        return scrollbarPositionToOffset(pos, length, total);
+    }
+
+    private double breadthPositionToPixels(double pos) {
+        double total = totalBreadthEstimateProperty().getValue();
+        double breadth = sizeTracker.getViewportBreadth();
+        return scrollbarPositionToOffset(pos, breadth, total);
+    }
+
+    private static double offsetToScrollbarPosition(
+            double contentOffset, double viewportSize, double contentSize) {
+        return contentSize > viewportSize
+                ? contentOffset / (contentSize - viewportSize) * contentSize
+                : 0;
+    }
+
+    private static double scrollbarPositionToOffset(
+            double scrollbarPos, double viewportSize, double contentSize) {
+        return contentSize > viewportSize
+                ? scrollbarPos / contentSize * (contentSize - viewportSize)
+                : 0;
     }
 }
